Tidy fixes

+++ src/shared/InputLine.cpp
@@ -5,7 +5,7 @@ namespace shared {
 InputLine::InputLine(const std::string& text, int width, int aMaxLen)
     : TInputLine(
           TRect(0, 0, width + 1 /*why is +1 needed?*/, 1),
-          std::max(text.size() + 1, static_cast<size_t>(aMaxLen) + 1)) {
+          std::max(static_cast<int>(text.size()) + 1, aMaxLen + 1)) {
     memcpy(data, text.c_str(), text.size());
     data[text.size()] = 0;  // NOLINT(cppcoreguidelines-pro-bounds-pointer-arithmetic)
 }
